feat(all): send and display outdoor temp
The weather client creates the data packet and sends it.
Home state receives it and displays the outdoor temperature.
Public API for components now has the component name (lcd_clear()).
HAL is prefixed with _hal (lcd_hal_send_cmd()) to keep separation.

+++ host/src/weather/weather.c
@@ -22,7 +22,7 @@ int weather_get_temp(const char *raw_json, struct weather_data *data_out) {
         return WEATHER_E_JSON;
     }
 
-    data_out->temp = temp->valuedouble;
+    data_out->temp = (float) temp->valuedouble;
     
     cJSON_Delete(root);
     return WEATHER_OK;
